Update to Item class
