+++ data/book.json
@@ -1,0 +1,1 @@
+[{"title":"adnan","publisher":null,"cover_state":null,"publish_date":"good"}]
